Always pass proper parent to Xonomy.xml2js() function
There are cases when full element hierarchy (including
top-level parent) is necessary for proper element rendering.

Consider the following configuration:

  var docSpec = {
    elements: {
      item: {
        collapsed: function(jsElement) {
          return js.parent().name === 'document';
        }
      },
      ...
    }
  };

In such case it is crucial to construct jsElement including
the parent element, otherwise the user-defined function fails.

Please, note that this change can have a significant performance
hit of harvesting parent element even if it is not necessary.
Alternatively we could consider a lazy approach - modify
Xonomy.xml2js() to accept 'parent' parameter as a function.
Then instead of this:

  var jsParent = Xonomy.harvestElement(document.getElementById(htmlID));
  html+=Xonomy.renderElement(Xonomy.xml2js(xml, jsParent));

... we could call it as follows:

  function parent() {
    return Xonomy.harvestElement(document.getElementById(htmlID));
  };
  html+=Xonomy.renderElement(Xonomy.xml2js(xml, parent));